Enable Django debug toolbar
This commit enables the Django debug toolbar. Note that the toolbar
is only visible when the app is in DEBUG mode and when you're
coming from IP 127.0.0.1. One other thing worth noting is that
the changeset updates requirements.txt to install the debug
toolbar directly from git, because that version is compatible
with new-style Django MIDDLEWARE.