Move queryset aggregations to their own view/mixin
This commit creates two custom mix-ins that can be used together
to create aggregate endpoints. AggregateQuerysetMixin exposes
the aggregate logic that was previously in the AggregateHandler.py
and makes it available to any view. FilterQuerysetMixin encapsulates
the filtering logic (longer term goal would be to consider replacing
our custom filters with filter_queryset/django filters).
In addition, the commit adds an aggregate view that can be used
in conjunction with the filter mixin to create aggregate-based
endpoint (see AwardsListAggregate for an example).