Add total and page metadata to aggregate view response
Add a total_metadata and page_metadata to the response returned
by the AggregateView class. This is for consistency with the
responses returned by the rest of the API. Note that this
commit also removes the option for using the AggregateView
without a groupby field specified in the request. Having to
handle two types of data coming back (querysets vs the
dictionaries) was adding complexity for no real reason (e.g.,
we don't have a use case for someone wanting to aggregate
the entire universe of a model's data).