Remove endpoint-specific aggregate values and unique_values option
The team agreed that we should separate aggregate and detailed
data into separate endpoints. This commit updates the docs in
preparation for 1) the upcoming aggregate endpoint and 2) the
move to using generics/mixins for our detailed endpoints. As
we start refactoring those views, the unique values and
endpoint-specific aggregates will disappear.