Prototype the viewset/mixin approach to endpoint view
Using the awards/summary endpoint, test using a combination
of generic viewset and mixins for our detailed data view
instead of inheriting from APIView. Making this switch lays
the groundwork for moving towards some of the functionality
that comes out of the Django Rest Framework box. For example,
accessing the request directly from a serializer, using DRF's
built-in pagination to return browsable URLs, and using a
combo of viewsets/routers to make it easier to wire up the
URLs when jumping between a list of model objects and a
single object.