Retain QueryDict object when parsing request parameters
The code was casting the query_param and data parts of the
incoming request object from a QueryDict to a Dict, which
caused the data on a GET request to show up as a list. Better
to stick with the intended behavior and keep everything in a
QueryDict so it can be properly parsed.