Switch award views to viewsets
This changeset switches the award list view to the custom
ViewSet we're using for detailed data endpoints. It also
switches the award list summary endpoint from DRF's generic
ViewSet to the custom viewset (aka DetailViewSet). What's
left to do here is to switch AwardListAggregate to a ViewSet
instead of a generic view to be consistent with how we construct
our endpoints. For now, we're ignoring the AwardListSummaryAutocomplete
view (it still works, it's just not inheriting from a ViewSet).