Convert custom AggregateView to inherit from ViewSet instead of view
For consistency with the other views, switch common.views.AggregateView
to inherit from a ViewSet instead of View. This involved tweaking
the associated code in urls.py too. As part of this process, I
found that we didn't have an integration test on the award/totals
endpoint, so I added a basic one to ensure that future url-related
updates don't break the endpoint. We still need some robust
unit tests on all of the endpoints.