Move award lookup and creation logic to Award model
This commit moves award lookup and creation logic from a
function in the submission loader to a method on the Award
model, so it will be accessible to both the submission loader
and the usaspending historic data loader. Note that a minor
change here is switching the order in which fain and uri are
processed to ensure that fain takes priority when both are
specified on an award transaction record.