Add funding agency to award transactions
Since one of our acceptance criteria for creating award records is to
include the funding agency, this field needs to be capture when
we're loading award transactions. The information is available on
the source data, so it should just be a matter of updating the
schema and then updating the load process (which will happen
in a subsequent commit).