Add place of performance to AwardAction
Because we've had some recent conversations about the types of
award data than can change from transaction to transaction, decided
to err on the side of retaining more information at the transaction
level and just using the most recent information when creating the
corresponding award. This also will help streamline the award
creation/update process, since we won't need to capture place of
performance when creating the shell award but can instead rely on
the Award models' update_from_mod method to fill this in.