Do the AwardAction schema and data migrations separately
When running the data and migrations as a single entity, we'd
sometimes get a postgres error that failed the migrations due
to a trigger action being in a pending state. This splits the
work into three steps: adding the new fields, populating the
new fields based on values in the old fields, and then removing
the old fields.