Convert test_accounts.py to pytest fixtures/syntax
This commit removes the Django TestCase class from tests_accounts.py.
Originally, this change was made to help test an issue with test
db setup. Although the solution to that problem ultimately didn't
rely on the pytest conversion, I'm leaving it here since an unrelated
goal is to move all of our tests to pytest format.