Add transactions endpoint
This endpoint is what we'll aggregate on for the spending over time
graph (the aggregate piece will be added in the next commit). For
now, we're not actually using this endpoint for anything other than
the basis of the aggregation, so it is based on the procurement
model only and attempts to mock how the response will look once
we convert the AwardAction abstract model to a non-abstract
Transaction model.