Add endpoint to retrieve info about a single award
This extends the work that Catherine did to map an award
endpoint url with a primary key (/awards/1234/) to the same viewset
we're using to display a list of awards. Currently, the single
award response is still serving the same default fields that we use
for the award list, so the next step will be to fix that and
return all fields when requesting a single award.