Switch order of GET/POST check when creating filters
This commit handles an issue that occurs when an incoming
POST request has both query parameters and a request body.
Instead of assuming that the presence of query parameters means
that there is no information in the request body, check the body
first and if there is info there, use that to create filters.
This is a temporary stopgap until we can better integration our
filter creation with the patterns recommended by Django Rest
Framework.