Switch ViewSets to built-in Django Rest Framework pagination
This commit is the first step towards moving away from our
own pagination. There's a new pagination.py file that now
automatically adding page and total metadata to any response
that's rendered via one of our ViewSets. For now, the response
mimics the format of our previous response, though in the future
we could likely switch to something flatter. Two things of
special note: 1 - DetailViewSet is now a shell, since we no
longer need the list override to generate a custom response. It
remains in the code because future for the single item endpoints
will it. 2 - DRF's pagination requires that the 'page' and 'limit'
options are specified as query params. Thus, we will no longer be
able to add 'page' and 'limit' in the request content body. Updated
tests and docs to follow.