Add page and limit options back to POST requesets temporarily
After consulting with the front-end team, we decided to keep
the page and limit post options intact until they have time to
update their code. The bulk of the pagination changes can remain
in the PR. We'll just need to remove our custom DetailViewSet
once we're ready to make the switch.