Remove valid field check from aggregate mixin
It turns out that the aggregate code supported Django fk
traversal all along. WHat was stopping was a parameter validation
that flagged fk notation as an invalid field and aborted the request.
Adding all of those checks seemed like a good idea at the time, but
in retrospect, I think it's fine to let Django itself tell us when
the field is invalid. If we want to make that message nicer from
a user perspective, we should handle at a higher level in the app.