Add award dependency to location data migration
Tests were failing because their data migrations were running in a
different order--reference migrations ran before awards migrations,
which caused an error on the location data migration ('award table
doesn't exist'). Although it's generally a good idea to update
an old migration, putting in a dependency on an arbitrary award
migration seemed to fix the problem.