Re-add Brazil to the test fixtures, since we have a Brazilian test record
We didn't notice this before because the threaded loader is wrapped in
a try/catch. Testing w/o the threaded loader resulted in a test
failure because the country wasn't found in the db. I made a note
that we should find out what should happen when we get an unknown
country in the data.