Create ETL app to house all data-loading logic
This commit creates an ETL app for data load logic. Mostly, this
is convenient as we're transitioning from the AwardAction model
to the new Transaction models, since it allows the old loaders
to run for the duration. Also, now that we're further along, it's
becoming clear that ETL spans across all apps, so it might be
more straightfoward to keep it all in the same place, rather
than having some in awards, some in submissions, etc. This
changeset converts the contract loader to the new models
(financial assistance and submissions to come). Note that until
we move everything and do some cleanup, there will be some
duplicate code. Lastly, a note about the threaded loader: this
version of the contract loader doesn't use it because it was
set up to load one model at a time, and the Transaction/
TransactionContracts/TransactionAssistance models are designed
to by loaded together--one pair for each incoming .csv row.