Update award_update helpers to optionally update specific awards
This PR follows up on a to do and a code review suggestion. Because
award submission_ids are themselves modified as part of the
award update process, we're not using that to hone in on the
awards that need updated. Instead we pass in a tuple of award
ids that should be updated.