Convert endpoints and tests from AwardActions to Transaction
This commit transitions the transaction endpoint to the new
set of transaction tables. In addition, the old procurement_set
and financial_assistance_set information in the awards endpoing
has been updated to reflect the new model structure. We'll want
to think more about how much transaction info we want to return
in the awards endpoint (vs the specific transaction endpoint).
All tests and data now reflect the new models, so a future
commit will remove the old models.