Use Django migration instead of SQL for AwardAction cleanup
This is definitely not...great to fix up a migration after the fact.
I think things originally went awry when I tried to do some physical
table cleanup before removing the django abstract model that referenced
said tables. When I finally removed the django piece, the automigrations
tried to delete the tables again, causing an error. If anyone has a
better idea of a way to fix this, am happy to hear it!