Fixing Full Text Search
While helping @ejgullo on her full text search story, I discovered that
we were missing the code that performed vector annotation. It appears
that during refactoring to the new mixins, the annotation of full-text
search vectors was left out. This code rectifies this, cleans up the
filter mixin, adds validation for full text search operation, and
introduces a test case to catch this problem.

* Added extra validations for full text search queries to the filter
validator
* Added search vector annotation to the filtering mixin
* Removed deprecated code from the filtering mixin
* Added test case for full text search on multiple fields