Override viewset retrieve method to apply caching decorator
The front-end team noticed that performance on individual
awards was still very poor. It turns out that those urls
route to the viewset retrieve method (which makes sense). Our
code doesn't override the retireve method, thus we weren't
applying the caching decorator to. This fixes that. There's
probably a more elegant way to do this, but let's see if
this helps for now.