Add ObjectClass model to references app
ObjectClass model replaces RefObjectClassCode. I decided to wholesale
replace the existing object class table because its primary key was
the actual object class and not a surrogate key, which is not
consistent with how PK data works on the API's other endpoints. Since
we have to change the primary key, which means changing the submission
loader and doing a migration on existing data, decided take the
opportunity to rename the table. This changeset also adds the major
object class code and name that we need for the new OC endpoint and
split out the direct/reimbursable digit into a separate column to
ensure that we're storing info at its most atomic level.