Migrate FinancialAccountsByAwards to new ObjectClass model
Change FinancialAccountsByAwards object_class field to a FK
that references the new ObjectClass model and migrate the data.
This is a messy, multi-part data migration. Because of the way
that the original object_class column was named (it was overriding
Django's usual _id notation), the built-in migration operations
weren't working as expected. Thus, there are several raw SQL
statements in this series of migrations (renaming the original
object_class column to _old, doing a SQL update to get the id
from the new object class table, and eventually dropping the _old
column).

The migrations do irreversible data changes. That said, I concluded
that the time saved by doing the migration this way + the low risk
since these data pertain only to broker submissions was worth it.

Also fix up program activity