Re-tuning the API Tutorial to be more accessible to beginners.
Explains who the tutorial is for and directs folks with more API
experience (defined) to "Using the API." Directs folks who want
human-readable versions of the data to usaspending.gov.

Better defines terms throughout.

Changes to the header structure for 508-compliance and readability
(removing multiple H1s and re-nesting headers for ease of
navigation).