Add quarterly data method to FinancialAccountsByProgramActivityObjectClass
This PR adds a class method to FinancialAccountsByProgramActivityObjectClass
that returns a RawQuerySet to return the quarterly breakdown of
financial numbers by tas, object class, and program activity. Using
Django's .raw() is a raw sql compromise. Subsequent processes that
call this method to load a quarterly table and whatnot can still access
the returned data via FinancialAccountsByProgramActivityObjectClass object
constructs, so we're not entirely cut off from the ORM.

Add ability to get qtrly numbers for specific submission