Add function and tests to find the most recent previous submisison in the FY
This PR creates get_previous_submission, which uses a cgac_code, fiscal
year, and fiscal period to see if if usaspending has any previous
broker submissions loaded for fiscal year. We'll call this function
in the submission loader to ensure that the new previous_submission
field is populated when creating a SubmissionAttribute record.