Update submission loader to find and populate previous_submission
This changeset updates the submission loader to look find a
previous submission when loading a new submission from the broker.
From the submisison being loaded, we'll use the cgac code, fiscal
year, and fiscal period to find a previous submission and use it
to populate the new submissions previous_submission field.

Also move the submission creation logic to its own function
for ease of unit testing.

Note that chaining submissions together like this supports the
work to calculate individual quarterly submission data. However,
it does introduce a complication when deleting submissions Until
we get clarification, the submission loader will not allow you
to delete a submission if it's the "previous" submission for
something downstream.