Remove everything from financial_activities app except models and tests
We now serve tas/object class/program activity (aka "File B")
breakouts via the tas/categories endpoint, which lives in the accounts
application. It was a bit confusing that the corresponding serializer
and view were still in the old financial_activities application, so
this commit moves those to the accounts app. In addition, it removes
the no-longer-used financial_activities url. Eventually we should
move the financial_activities models to the accounts app and remove
financial_activities altogether, but that's not trivial, and this
isn't the time.