Add object class, program activity, and tas serialization to TAS categories
Update TasCategorySerializer to include nested serializers for program
activity, object class, and tas. Note that the tas serializer needs
a bit more work to ensure that its "verbose" version apply here, but
will leave it for now and circle back after we merge the first pass
of the endpoint updates.

Note that the commit also removes some of the WIP for the abandoned-for-now
"fancy" version of the tas endpoint.