Force migration to run after dependencies in other apps are complete
This addresses another instance where my migrations suddenly begin
to run out of order when running the test suite locally. The problem
usually involves migration that drops a model running prior to a migration
in another app that references said model. Longer term, we need to
understand why this is happening, but in the short term, I needed to
fix local test setup.