Load 2017 program activity data
This commit removes program activity reference data from our fixtures and
loads it from a .csv file instead. Unlike some of the other loaders, it
doesn't make use of the threaded loader because that doesn't support keying
on multiple columns. The data is pretty straightforward, and wrapping the
load into a single transaction results in reasonable performance.