Do not cascade delete the awards table when submission is deleted
While we rely on a series of deleting cascades for the submission
delete process to work, we should not cascade delete the award
table because it's possible that records unrelated to the
submission being deleted point to them.

At some point, we'll need a smarter delete process to ensure
that we don't leave orphaned award records lying around, but
for now, we'll at least breaking existing links.