Make the broker submission loader delete an existing submission by default
Previously, we required an explicit --delete flag on the load_submission
command to clear out a previous load when a broker submission gets reloaded.
Since we never really had a mechanism to update existing records, we decided
to eliminate that as an option and have the loader delete existing submissions
by default.

This changeset also uses the new rm_submission command when deleting
an existing submission, so now we  have all of our submission
delete logic in a single place.