Fix missing awarding/funding agency cgac codes on incoming broker D1/D2
The process that generates the broker's D1 and D2 files (i.e.,
award transaction records) is supposed to fill in the CGAC code of
the awarding and funding top-tier agencies. In the event these
fields are still empty when we get them, use the subtier agency
codes to look up the toptier/cgac info so we can load it.