Change CharFields in models to TextField
We made a decision to globally use TextField instead of CharField in
the API data models. Because data validation happens in the broker,
it's not the role of the API to validate it via these types of
db constraints. We don't want to be tightly coupled to the data
model in that way, and for the first round of loads, we know
that some upstream fields may be longer than anticipated, so we
want to avoid failed jobs related to that.