Fix tests to accommodate the inability to fill optional Location fields
The location model has a long unique index that cannot be created for
records that have large numbers of characters in all of the model's
optional fields. In reality, this scenario is unlikely to occur.
In unit testing, however, it means that the model mommy
 _fill_optional param can't be used. We should almost certainly
explore a different type of index in the future, but for now, I
think it's ok to err on the side of guess that most incoming locations
will be sparsely populated enough not to trigger the index error (versus
keeping character limitations that risk a load error when incoming data
exceeds it).