Update submission loader to handle duplicate object classes in File B
This changeset adds load logic to inspect an incoming File B for duplicate
4 digit object class codes. This is happening for two agency submissions
who are using 4 digit object codes with different D/R flags to report
deobligated funds. This will be fixed on the schema/broker end in future
reporting periods, but for now, we need to load the certified
submissions.

In theory, we could use the same "aggregate" query to pull File B
data, regardless of whether or not the submission has this particular
object class quirk. That would make the code a bit cleaner, but I'd
rather err on the side of "do not attempt a fancy query unless we
really have to."