Reset database migrations
This changeset removes the existing migration .py files for our Django apps.
It adds a new series of 0001_initial migrations that reflect the current state
of the database. Lastly, it adds a management command that removes the existing
migration record from the django_migrations table and runs fake migrations for
each app. Once the management command (reset_migrations) is run, we'll be back
at ground zero for migrations.