explorers
agency, and budget_function explorers seperated out into packages.

Since there is a lot of overlap in the various types and not all types are using the same parameters to build or filter on, this seem to be the best approach for now.

{"type": "budget_function"} etc, limits results when attempting to multiple starting points. Since subsequent parameters change depending on the starting point.

For agency - { "agency_type": with one of the following ['agency', 'federal_account', 'program_activity', 'object_class', 'recipients', 'awards'] }

For budget_function - {
"budget_type": with one of the following ['budget_function', 'budget_subfunction', 'federal_account', 'program_activity', 'object_class', 'recipients', 'awards']