[DEV-6756] Allowing version conflicts on deletes to be ignored, and not exiting but raising exception
In the upfront DB delete sync, it is just doing deletes and no other changes on the index, so any conflict will be a conflict with another delete of the same doc = ignore. In the batched deletes, each partition-batch being deleted should be a unique doc being updated or deleted, not concurrent modification/deletes from other batches.